backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-environ==0.12.0
psycopg2-binary==2.9.11
requests==2.32.3
